requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
fastapi==0.78.0
fonttools==4.33.3
h11==0.13.0
httptools==0.4.0
idna==3.3
imageio==2.19.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
pip==21.3.1
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
setuptools==60.2.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
tifffile==2022.5.4
typing-extensions==4.2.0
uvicorn==0.18.2
watchfiles==0.15.0
websockets==10.3
wheel==0.37.1
python-multipart